fix: code editable attribute was sometimes missing
race conditions since the migration to the store
